v0.041

+++ app/src/main/java/com/xmum/hiyapodcast/interfaces/IPlayerCallback.java
@@ -1,25 +1,38 @@
 package com.xmum.hiyapodcast.interfaces;
 
-
-
 import com.ximalaya.ting.android.opensdk.model.track.Track;
 import com.ximalaya.ting.android.opensdk.player.service.XmPlayListControl;
 
 import java.util.List;
 
 public interface IPlayerCallback  {
+
     void onPlayStart();
+
     void onPlayPause();
+
     void onPlayStop();
+
     void onPlayError();
+
     void nextPlay(Track track);
+
     void onPrePlay(Track track);
+
+    //play list has loaded
     void onListLoaded(List<Track> list);
+
+    //change play mode
     void onPlayModeChange(XmPlayListControl.PlayMode playMode);
-    void onProgressChange(long currentProgress,long total);
+
+    //change the progress bar
+    void onProgressChange(int currentProgress,int total);
+
+    //load the ad from sdk
     void onAdLoading();
+
+    //ad load finished
     void onAdFinished();
 
-    
 
 }
